requirements.txt
fastapi
uvicorn
python-multipart
openai
python-docx
pandas
aiofiles
